rename ToolButton::menu_item to action_name. Don't mark action names for
2006-05-25  Hans Breuer  <hans@breuer.org>

	* app/interface.[hc] app/menus.c : rename ToolButton::menu_item
	to action_name. Don't mark action names for translation. It is not
	needed and some languages don't support CamelCase. Fixes bug #342895

	* app/menus.c : re-added a bunch of accelerators lost in translation
	to the GtkAction framework. Use GTK_STOCK_FULLSCREEN conditionally
	it was not available with gtk+-2-6-x.
	Use dia_get_data_directory() instead of hardcoded UIDATADIR.
	* app/dia_embedd.c : explain instead of out-comment

	* app/menu.c lib/autoroute.c lib/color.c lib/diagramdata.c
	  lib/properties.h dia_xml.c : warningectomy

	* lib/create.h : create_standard_group()
	* lib/libdia.def lib/makefile.msc : reflect moved stuff

+++ lib/color.c
@@ -76,9 +76,9 @@ color_new_rgb(float r, float g, float b)
 void
 color_convert(Color *color, GdkColor *gdkcolor)
 {
-  gdkcolor->red = color->red*65535;
-  gdkcolor->green = color->green*65535;
-  gdkcolor->blue = color->blue*65535;
+  gdkcolor->red = (guint16)(color->red*65535);
+  gdkcolor->green = (guint16)(color->green*65535);
+  gdkcolor->blue = (guint16)(color->blue*65535);
 
   if (_color_initialized) {
     if (!gdk_colormap_alloc_color (colormap, gdkcolor, TRUE, TRUE))
